refactor: apply Open-Closed Principle to ExecuteTimestampQuery for extensibility
