+++ src/Delta/TimestampQueries/ITimestampQueryExecutor.cs
@@ -1,0 +1,6 @@
+namespace Delta;
+
+interface ITimestampQueryExecutor
+{
+    Task<string> Execute(DbCommand command, Cancel cancel);
+}
